# python-flask-server/swagger_server/models/matricula.py
# coding: utf-8

from __future__ import absolute_import
from .base_model_ import Model
from datetime import date, datetime
from typing import List, Dict
from ..util import deserialize_model


class Matricula(Model):
    """
    NOTE: This class is auto generated by the swagger code generator program.
    Do not edit the class manually.
    """
    def __init__(self, asignaturas: str=None):
        """
        Matricula - a model defined in Swagger

        :param asignaturas: The asignaturas of this Matricula.
        :type asignaturas: str
        """
        self.swagger_types = {
            'asignaturas': str
        }

        self.attribute_map = {
            'asignaturas': 'asignaturas'
        }

        self._asignaturas = asignaturas

    @classmethod
    def from_dict(cls, dikt) -> 'Matricula':
        """
        Returns the dict as a model

        :param dikt: A dict.
        :type: dict
        :return: The Matricula of this Matricula.
        :rtype: Matricula
        """
        return deserialize_model(dikt, cls)

    @property
    def asignaturas(self) -> str:
        """
        Gets the asignaturas of this Matricula.
        asignaturas en las que se va a matricular

        :return: The asignaturas of this Matricula.
        :rtype: str
        """
        return self._asignaturas

    @asignaturas.setter
    def asignaturas(self, asignaturas: str):
        """
        Sets the asignaturas of this Matricula.
        asignaturas en las que se va a matricular

        :param asignaturas: The asignaturas of this Matricula.
        :type asignaturas: str
        """

        self._asignaturas = asignaturas