# python-flask-server/swagger_server/swagger/swagger.yaml
...
swagger: "2.0"
info:
  description: "Aplicación para gestionar los datos de la universidad."
  version: "2.0.0"
  title: "Universidad de Villamayor"
  termsOfService: ""
...
schemes:
- "https"
paths:
...
  /Alumnos:
    get:
      tags:
      - "Alumnos"
      summary: "Consultar los datos de un alumno"
      operationId: "get_datos"
      produces:
      - "application/xml"
      - "application/json"
      parameters:
      - name: "dni"
        in: "query"
        description: "Da informacion acerca de un alumno"
        required: true
        type: "string"
        items:
          type: "array"
        collectionFormat: "multi"
      responses:
        200:
          description: "successful operation"
        400:
          description: "Invalid Order"
      x-swagger-router-controller: "swagger_server.controllers.alumnos_controller"
    post:
      tags:
      - "Alumnos"
      summary: "crear un nuevo alumno"
      operationId: "crear_alumno"
      produces:
      - "application/xml"
      - "application/json"
      parameters:
      - in: "body"
        name: "alumno"
        description: "Crear Alumno"
        required: false
        schema:
          $ref: "#/definitions/Alumno"
      responses:
        200:
          description: "successful operation"
        405:
          description: "Validation exception"
      x-swagger-router-controller: "swagger_server.controllers.alumnos_controller"
    put:
      tags:
      - "Alumnos"
      summary: "Modificar alumno"
      operationId: "modificar_alumno"
      produces:
      - "application/xml"
      - "application/json"
      parameters:
      - in: "body"
        name: "dni"
        description: "dni del alumno que se quiere modificar"
        required: true
        schema:
          $ref: "#/definitions/Alumno"
      responses:
        200:
          description: "successful operation"
        400:
          description: "DNI inválido"
        404:
          description: "Alumno not encontrado"
        405:
          description: "Validation exception"
      x-swagger-router-controller: "swagger_server.controllers.alumnos_controller"
  /Alumnos/Matricula:
    get:
      tags:
      - "Alumnos"
      summary: "Consultar matricula"
      operationId: "get_matricula"
      produces:
      - "application/xml"
      - "application/json"
      parameters:
      - name: "dni"
        in: "query"
        description: "Muestra la información de la matrícula de un alumno"
        required: true
        type: "string"
        items:
          type: "array"
        collectionFormat: "multi"
      responses:
        200:
          description: "successful operation"
        400:
          description: "Invalid Order"
      x-swagger-router-controller: "swagger_server.controllers.alumnos_controller"
    post:
      tags:
      - "Alumnos"
      summary: "Matricularse"
      operationId: "matriculacion"
      produces:
      - "application/xml"
      - "application/json"
      parameters:
      - in: "body"
        name: "matricula"
        description: "Crear Matricula"
        required: false
        schema:
          $ref: "#/definitions/Matricula"
      responses:
        200:
          description: "successful operation"
        405:
          description: "Validation exception"
      x-swagger-router-controller: "swagger_server.controllers.alumnos_controller"
...
securityDefinitions:
  api_key:
    type: "apiKey"
    name: "api_key"
    in: "header"
definitions:
...
  Alumno:
    type: "object"
    properties:
      NombreCompleto:
        type: "string"
      DNI:
        type: "string"
      FechaNacimiento:
        type: "string"
      Grado:
        type: "integer"
        format: "int64"
        description: "Código del grado que está cursando"
      Usuario:
        type: "string"
        description: "username del loggin"
      Contraseña:
        type: "string"
        description: "contraseña del loggin"
      email:
        type: "string"
        description: "email del alumno"
  Matricula:
    type: "object"
    properties:
      asignaturas:
        type: "string"
        description: "asignaturas en las que se va a matricular"
...